feat: improve settings layout and functionality with enhanced navigation and new settings page
