+++ packages/frontend/components/SettingsPanel.tsx
@@ -177,7 +177,7 @@ function SettingsPanel({ visible, onClose, onLogout }) {
           className="flex-row items-center mb-2 p-2 rounded-lg hover:bg-gray-100 dark:hover:bg-neutral-800"
           onPress={() => {
             onClose()
-            router.push('/settings')
+            router.push('/settings/settings')
           }}
         >
           <Settings size={16} color={isDark ? '#fff' : '#374151'} className="mr-3" />
